# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SanTech Data VTU Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /data/plans:
    get:
      operationId: getDataPlans
      tags: [data]
      summary: Get available data plans
      security:
        - bearerAuth: []
      parameters:
        - name: network
          in: query
          schema:
            type: string
            enum: [MTN, AIRTEL, GLO, 9MOBILE]
      responses:
        "200":
          description: Data plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DataPlan"
  /data/purchase:
    post:
      operationId: purchaseData
      tags: [data]
      summary: Purchase a data bundle
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DataPurchaseInput"
      responses:
        "200":
          description: Data purchased
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"

  # ─── AIRTIME ─────────────────────────────────────────────────────────────────
  /airtime/purchase:
    post:
      operationId: purchaseAirtime
      tags: [airtime]
      summary: Purchase airtime
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AirtimePurchaseInput"
      responses:
        "200":
          description: Airtime purchased
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"

  # ─── ELECTRICITY ─────────────────────────────────────────────────────────────
  /electricity/providers:
    get:
      operationId: getElectricityProviders
      tags: [electricity]
      summary: Get electricity providers (DISCOs)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Providers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ElectricityProvider"
  /electricity/verify-meter:
    post:
      operationId: verifyMeter
      tags: [electricity]
      summary: Verify meter number
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeterVerifyInput"
      responses:
        "200":
          description: Meter info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeterInfo"
  /electricity/purchase:
    post:
      operationId: purchaseElectricity
      tags: [electricity]
      summary: Purchase electricity token
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ElectricityPurchaseInput"
      responses:
        "200":
          description: Token purchased
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ElectricityTransaction"

  # ─── CABLE TV ────────────────────────────────────────────────────────────────
  /cable/providers:
    get:
      operationId: getCableProviders
      tags: [cable]
      summary: Get cable TV providers
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Cable providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CableProvider"
  /cable/plans:
    get:
      operationId: getCablePlans
      tags: [cable]
      summary: Get cable TV plans
      security:
        - bearerAuth: []
      parameters:
        - name: provider
          in: query
          schema:
            type: string
            enum: [DSTV, GOTV, STARTIMES]
      responses:
        "200":
          description: Cable plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CablePlan"
  /cable/verify-smartcard:
    post:
      operationId: verifySmartcard
      tags: [cable]
      summary: Verify smartcard/IUC number
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SmartcardVerifyInput"
      responses:
        "200":
          description: Smartcard info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmartcardInfo"
  /cable/subscribe:
    post:
      operationId: subscribeCable
      tags: [cable]
      summary: Subscribe to cable TV plan
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CableSubscribeInput"
      responses:
        "200":
          description: Subscription successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"

  # ─── EXAM TOKENS ─────────────────────────────────────────────────────────────
  /exam/types:
    get:
      operationId: getExamTypes
      tags: [exam]
      summary: Get available exam types
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Exam types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExamType"
  /exam/purchase:
    post:
      operationId: purchaseExamToken
      tags: [exam]
      summary: Purchase exam token/pin
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamPurchaseInput"
      responses:
        "200":
          description: Exam token purchased
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamTransaction"

  # ─── TRANSACTIONS ────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DataPlan:
      type: object
      required: [id, network, name, size, validity, price, isActive]
      properties:
        id:
          type: string
        network:
          type: string
          enum: [MTN, AIRTEL, GLO, 9MOBILE]
        name:
          type: string
        size:
          type: string
        validity:
          type: string
        price:
          type: number
        costPrice:
          type: ["number", "null"]
        isActive:
          type: boolean
...
    DataPurchaseInput:
      type: object
      required: [planId, phone]
      properties:
        planId:
          type: string
        phone:
          type: string

    # ─── AIRTIME SCHEMAS ──────────────────────────────────────────────────────
    AirtimePurchaseInput:
      type: object
      required: [network, phone, amount]
      properties:
        network:
          type: string
          enum: [MTN, AIRTEL, GLO, 9MOBILE]
        phone:
          type: string
        amount:
          type: number

    # ─── ELECTRICITY SCHEMAS ──────────────────────────────────────────────────
    ElectricityProvider:
      type: object
      required: [id, name, code]
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
    MeterVerifyInput:
      type: object
      required: [meterNumber, providerCode, meterType]
      properties:
        meterNumber:
          type: string
        providerCode:
          type: string
        meterType:
          type: string
          enum: [prepaid, postpaid]
    MeterInfo:
      type: object
      required: [meterNumber, name, address]
      properties:
        meterNumber:
          type: string
        name:
          type: string
        address:
          type: string
    ElectricityPurchaseInput:
      type: object
      required: [meterNumber, providerCode, meterType, amount, phone]
      properties:
        meterNumber:
          type: string
        providerCode:
          type: string
        meterType:
          type: string
          enum: [prepaid, postpaid]
        amount:
          type: number
        phone:
          type: string
    ElectricityTransaction:
      type: object
      required: [id, status, token, amount, meterNumber, createdAt]
      properties:
        id:
          type: string
        status:
          type: string
        token:
          type: ["string", "null"]
        amount:
          type: number
        meterNumber:
          type: string
        createdAt:
          type: string

    # ─── CABLE SCHEMAS ────────────────────────────────────────────────────────
    CableProvider:
      type: object
      required: [id, name, code]
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
    CablePlan:
      type: object
      required: [id, provider, name, price, validity]
      properties:
        id:
          type: string
        provider:
          type: string
          enum: [DSTV, GOTV, STARTIMES]
        name:
          type: string
        price:
          type: number
        validity:
          type: string
    SmartcardVerifyInput:
      type: object
      required: [smartcardNumber, provider]
      properties:
        smartcardNumber:
          type: string
        provider:
          type: string
          enum: [DSTV, GOTV, STARTIMES]
    SmartcardInfo:
      type: object
      required: [smartcardNumber, name, currentPlan]
      properties:
        smartcardNumber:
          type: string
        name:
          type: string
        currentPlan:
          type: string
        dueDate:
          type: ["string", "null"]
    CableSubscribeInput:
      type: object
      required: [planId, smartcardNumber, provider]
      properties:
        planId:
          type: string
        smartcardNumber:
          type: string
        provider:
          type: string
          enum: [DSTV, GOTV, STARTIMES]

    # ─── EXAM SCHEMAS ─────────────────────────────────────────────────────────
    ExamType:
      type: object
      required: [id, name, code, price]
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
          enum: [WAEC, NECO, JAMB, NABTEB]
        price:
          type: number
        description:
          type: ["string", "null"]
    ExamPurchaseInput:
      type: object
      required: [examTypeId, quantity, phone]
      properties:
        examTypeId:
          type: string
        quantity:
          type: integer
        phone:
          type: string
    ExamTransaction:
      type: object
      required: [id, status, pins, examType, amount, createdAt]
      properties:
        id:
          type: string
        status:
          type: string
        pins:
          type: array
          items:
            type: object
            required: [pin, serial]
            properties:
              pin:
                type: string
              serial:
                type: string
        examType:
          type: string
        amount:
          type: number
        createdAt:
          type: string

    # ─── TRANSACTION SCHEMAS ──────────────────────────────────────────────────
    Transaction:
      type: object
      required: [id, type, status, amount, description, createdAt]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [data, airtime, electricity, cable, exam, wallet_fund, wallet_transfer]
        status:
          type: string
          enum: [pending, success, failed]
        amount:
          type: number
        description:
          type: string
        reference:
          type: ["string", "null"]
        metadata:
          type: object
          additionalProperties: true
        userId:
          type: string
        createdAt:
          type: string
...